Bubo/requirements.txt
SpeechRecognition
PyAudio
weather-api